+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix #111
 - Fix #112
 - Add restorationScopeId for the routers and auto restoration
+- add global middlewares
 
 ## 1.8.1  25.02.2023